Improve resilience of requires WRT to directory layout.
It is not needed anymore to create a symlink to the Job/
directory.  You can pull p5-Job module directly here and
parallel.pl will be it automatically.

Also shut a warning is sys/ioctl.ph cannot be found and
do not try to use the ioctl in that case.