Handle half line independently for each filehandle.
When there are multiple commands executed remotely which produce
output on the same line, we may receive the line in multiple
reads and the previous algorithm splittd the line in order to not
lose the content of the "half line" buffer.  By having a buffer
per filehandle, we keep the half-line buffer across multiple
select(2).